+++ engine/executor.py
@@ -22,6 +22,8 @@ from .models import TestExecutor
 from .selenium_helper import *
 from .testreport_generator import *
 from .email_generator import *
+from daksha import settings
+from engine import test_result_utils
 
 import jinja2
 import ast
@@ -64,12 +66,16 @@ def execute_test(test_executor: TestExecutor, email):
             if execution_result is False:
                 break
         logger.info("Test " + name + " finished, generating  result ")
-        generate_result(test_executor.test_id, execution_result, name, step, error_stack)
+        generate_result(test_executor.test_uuid, execution_result, name, step, error_stack )
+        
+        if settings.TEST_RESULT_DB != None and settings.TEST_RESULT_DB.lower() == "postgres":
+            test_result_utils.save_result_in_db(test_executor,execution_result,step,error_stack)
+            
         if execution_result:
             logger.info("Test " + name + " successful")
         else:
-            logger.info("Test " + name + " failed for test ID: " + test_executor.test_id)
-            send_alert(test_executor.test_id, name, str(step), error_stack, alert_type)
+            logger.info("Test " + name + " failed for test ID: " + test_executor.test_uuid)
+            send_alert(test_executor.test_uuid, name, str(step), error_stack, alert_type)
         __cleanup(web_driver)
     except Exception:
         logger.error("Error encountered in executor: ", exc_info=True)