+++ engine/models.py
@@ -21,12 +21,13 @@ from django.db import models
 
 # Create your models here.
 class TestExecutor:
-    def __init__(self, index, test_id, variable_dict, test_yml, web_driver):
+    def __init__(self, index, test_uuid, variable_dict, test_yml, web_driver , test_result):
         self.index = index
-        self.test_id = test_id
+        self.test_uuid = test_uuid
         self.variable_dictionary = copy.deepcopy(variable_dict)
         self.test_yml = test_yml
         self.web_driver = web_driver
+        self.test_result=test_result
 
 
 class TestResult:
@@ -35,3 +36,13 @@ class TestResult:
         self.test_status = test_status
         self.failed_step = step
         self.failure_cause = failure_reason
+
+
+class TestResults(models.Model):
+    TestUUID=models.TextField(max_length=11)
+    TestName=models.TextField()
+    Status=models.TextField()
+    FailureStep=models.TextField()
+    FailureCause=models.TextField()
+    InsertTs = models.DateTimeField(auto_now_add=True) #The values updated here are following UTC time zone
+    UpdateTs = models.DateTimeField(auto_now=True) #The values updated here are following UTC time zone