+++ engine/thread_executor.py
@@ -21,20 +21,28 @@ from .logs import *
 from daksha.settings import APACHE_URL
 from .models import TestExecutor
 from .testreport_generator import generate_report
+from engine import test_result_utils
 
+def thread_executor(test_ymls, initial_variable_dictionary, test_uuid, email):
+    # Test Executor object initialization and store in a list
+    testExecutorObjects=[]
 
-def thread_executor(test_ymls, initial_variable_dictionary, test_id, email):
+    for test_yml in test_ymls:
+        test_result_object=test_result_utils.initialize_test_result(test_uuid,test_yml) 
+        test_executor = TestExecutor(1, test_uuid, initial_variable_dictionary, test_yml, None ,test_result_object)
+        testExecutorObjects.append(test_executor)
+           
+        
     with ThreadPoolExecutor(max_workers=3) as pool_executor:
-        for test_yml in test_ymls:
+        for test_executor in testExecutorObjects:
             try:
-                test_executor = TestExecutor(1, test_id, initial_variable_dictionary, test_yml, None)
                 pool_executor.submit(execute_test, test_executor, email)
                 logger.info("Task submitted")
             except Exception as e:
                 logger.error("Exception occurred", e)
             pass
     logger.info("All threads complete, generating test report")
-    generate_report(test_id)
-    report_url = APACHE_URL + test_executor.test_id + '/report.html'
-    send_report_email(test_executor.test_id, report_url, email)
+    generate_report(test_uuid)
+    report_url = APACHE_URL + test_executor.test_uuid + '/report.html'
+    send_report_email(test_executor.test_uuid, report_url, email)