+++ engine/views.py
@@ -37,22 +37,22 @@ def executor(request):
     """
     if request.method == 'POST':
         try:
-            test_id = generate_test_id()
-            os.makedirs(f"{STORAGE_PATH}/{test_id}")
-            logger.info('Directory created at: ' + f"{STORAGE_PATH}/{test_id}")
+            test_uuid = generate_test_uuid()
+            os.makedirs(f"{STORAGE_PATH}/{test_uuid}")
+            logger.info('Directory created at: ' + f"{STORAGE_PATH}/{test_uuid}")
             received_json_data = json.loads(request.body.decode())
             try:
-                test_ymls, initial_variable_dictionary = __extract_test_data(test_id, received_json_data['test'])
+                test_ymls, initial_variable_dictionary = __extract_test_data(test_uuid, received_json_data['test'])
             except BadArgumentsError as e:
                 return HttpResponse(str(e), status=status.HTTP_400_BAD_REQUEST)
             pool_executor = ThreadPoolExecutor(max_workers=1)
             try:
-                pool_executor.submit(thread_executor, test_ymls, initial_variable_dictionary, test_id,
+                pool_executor.submit(thread_executor, test_ymls, initial_variable_dictionary, test_uuid,
                                      received_json_data['email'])
                 logger.info("task submitted to thread pool executor")
             except Exception as e:
                 logger.error("Exception occurred", e)
-            response_message = "Your test ID is: " + test_id + ". We'll send you an email with report shortly"
+            response_message = "Your Test UUID is: " + test_uuid + ". We'll send you an email with report shortly"
             return HttpResponse(response_message, status=status.HTTP_200_OK)
         except Exception as e:
             logger.error("Exception caught", exc_info=True)
@@ -62,7 +62,7 @@ def executor(request):
         return HttpResponse(status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
 
-def __extract_test_data(test_id, test):
+def __extract_test_data(test_uuid, test):
     initial_variable_dictionary = {}
     if "variables" in test:
         initial_variable_dictionary = test['variables']
@@ -76,7 +76,7 @@ def __extract_test_data(test_id, test):
         else:
             files = get_yml_files_in_folder_git(REPO_NAME, BRANCH_NAME, path)
         for file_path in files:
-            test_yml = read_yaml(REPO_NAME, BRANCH_NAME, file_path, test_id)
+            test_yml = read_yaml(REPO_NAME, BRANCH_NAME, file_path, test_uuid)
             test_ymls.append(test_yml)
 
     elif "local" in source_location.lower():