+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # `DAKSHA` Change Log
 
-*version*: `2.1.3`
+*version*: `2.2.0`
+
+## v 2.2.0
+1. Added Database for Test Results in Daksha.
+2. Users can opt for test results to be saved in a Postgres database
 
 ## v 2.1.3
 1. Adressed issue #47